requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
idna==3.10
importlib_metadata==8.5.0
itchat-uos-fix==1.5.1.dev0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pypng==0.20220715.0
PyQRCode==1.2.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
